refactor: split canvas dragging and fitView features to individual composable functions
