+++ frontend/src/components/SchemaDiagram/Canvas/composables/index.ts
@@ -1,0 +1,2 @@
+export * from "./useDragCanvas";
+export * from "./useFitView";
