+++ frontend/src/components/SchemaDiagram/Canvas/const.ts
@@ -1,0 +1,4 @@
+export const ZOOM_RANGE = {
+  max: 2, // 200%
+  min: 0.05, // 5%
+};
