+++ frontend/src/components/SchemaDiagram/Navigator/index.ts
@@ -1,0 +1,3 @@
+import Navigator from "./Navigator.vue";
+
+export default Navigator;
